Fix few bugs
Fix issue with players not correctly respawning to where there pearl is
when released when they should be.
Fix returning summoning player being insta killed.
If a player has a bed they will be teleported to that on free if going
to their pearl isnt set to true in config.
If no bed is found they will random spawn to a random location on the
pearl server.
When returning a summoned player increase their spawn y by 1 so they
dont fall through the ground.